Add files via upload
This assignment uses the driver exam interface to access the DriverExam class which contains an array of correct answers and methods to check whether a supplied array of answers matches the correct answers. If not a complete match, there are methods to return the questions answered incorrectly & the number of correct & incorrect answers. There is no actual exam, as the programs are just for the purpose of demonstration.